Updating cli
