+++ proto/konstellation/oracle/genesis.proto
@@ -1,7 +1,8 @@
 syntax = "proto3";
 package types;
 
-// GenesisState defines the oracle module's genesis state.
+option go_package = "x/oracle/types";
+
 message GenesisState {
   string allowed_address = 1;
 }
